Corrigindo os arquivos de controle do PWM
Alterando o cabeçalho da função 'get_duty_cycle'.

Removendo o casting de tipo da função 'get_duty_cycle'.

Corrigindo o parâmetro da função 'pwm_set_gpio_level'.

+++ src/pwm.h
@@ -5,6 +5,6 @@
 
 void setup_pwm(uint pin,float pwm_clock_divisor, uint wrap, uint duty_cycle);
 
-uint get_duty_cycle(uint wrap, float percentage);
+float get_duty_cycle(uint wrap, float percentage);
 
 #endif
